refactor: enhance subscription management UI and remove unused components
