+++ apps/web/app/(settings)/settings/page.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useMemo, useRef, useState, type ChangeEvent } from "react";
 import { CardTitle } from "@/components/ui/card";
-import { Button } from "@/components/ui/button";
+import { Button, buttonVariants } from "@/components/ui/button";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Input } from "@/components/ui/input";
 import {
@@ -21,13 +21,191 @@ import { useAction, useMutation } from "convex/react";
 import { api } from "@teak/convex";
 import { authClient } from "@/lib/auth-client";
 import { Spinner } from "@/components/ui/spinner";
-import { AlertCircle, AlertTriangle, Monitor, Moon, Sun } from "lucide-react";
+import {
+  AlertCircle,
+  AlertTriangle,
+  Monitor,
+  Moon,
+  Sun,
+  CheckCircle2,
+  ArrowRight,
+  ExternalLink,
+} from "lucide-react";
 import { useRouter } from "next/navigation";
 import { Badge } from "@/components/ui/badge";
 import { useTheme } from "next-themes";
-import { SubscriptionSection } from "@/components/subscription-section";
 import Link from "next/link";
 import { Separator } from "@/components/ui/separator";
+import { PolarEmbedCheckout } from "@polar-sh/checkout/embed";
+import { cn } from "@/lib/utils";
+
+const featureList = [
+  "Unlimited Cards",
+  "Unlimited Storage",
+  "Automatic Summary and Tags",
+  "Automatic Audio Transcription",
+  "Chrome Extension",
+  "iOS Mobile App",
+  "Android Mobile App",
+];
+
+interface CustomerPortalButtonProps {
+  className?: string;
+  children: React.ReactNode;
+}
+
+function CustomerPortalButton({
+  className,
+  children,
+}: CustomerPortalButtonProps) {
+  const [isLoading, setIsLoading] = useState(false);
+  //@ts-ignore
+  const createCustomerPortal = useAction(api.billing.createCustomerPortal);
+
+  const handlePortal = async () => {
+    setIsLoading(true);
+    try {
+      const portalUrl = await createCustomerPortal({});
+      window.open(portalUrl, "_blank");
+      setIsLoading(false);
+    } catch (error) {
+      console.error("Failed to open customer portal", error);
+      toast.error("Failed to open customer portal. Please try again.");
+      setIsLoading(false);
+    }
+  };
+
+  return (
+    <button onClick={handlePortal} className={className} disabled={isLoading}>
+      {isLoading ? <Spinner /> : children}
+    </button>
+  );
+}
+
+interface PlanOptionProps {
+  planId: string;
+  title: string;
+  priceAmount: number;
+  intervalLabel: string;
+  badge?: string;
+  buttonVariant: "outline" | "destructive";
+  isLoading: boolean;
+  onCheckout: (planId: string) => void;
+}
+
+function PlanOption({
+  planId,
+  title,
+  priceAmount,
+  intervalLabel,
+  badge,
+  isLoading,
+  onCheckout,
+}: PlanOptionProps) {
+  const formattedPrice = priceAmount
+    ? `${(priceAmount / 100).toLocaleString()}$`
+    : "--";
+
+  return (
+    <div className="flex w-full flex-col justify-between rounded-md border bg-background p-5 text-left gap-4 relative overflow-hidden">
+      {badge && (
+        <Badge className="rounded-none absolute top-0 right-0 rounded-bl-md px-3">
+          {badge}
+        </Badge>
+      )}
+      <p className="font-medium text-muted-foreground">{title}</p>
+      <div className="flex justify-between items-end">
+        <div className="flex gap-2 items-end">
+          <p className="text-4xl font-semibold text-foreground">
+            {formattedPrice}
+          </p>
+          <p className="text-muted-foreground pb-1">{intervalLabel}</p>
+        </div>
+
+        <button
+          onClick={() => onCheckout(planId)}
+          className={cn(
+            buttonVariants({
+              variant: "outline",
+            })
+          )}
+          disabled={isLoading}
+        >
+          {isLoading ? (
+            <Spinner />
+          ) : (
+            <span className="flex items-center gap-2">
+              Continue <ArrowRight className="size-4" />
+            </span>
+          )}
+        </button>
+      </div>
+    </div>
+  );
+}
+
+interface SubscriptionSectionProps {
+  onCheckout: (planId: string) => void;
+  loadingPlanId: string | null;
+}
+
+function SubscriptionSection({
+  onCheckout,
+  loadingPlanId,
+}: SubscriptionSectionProps) {
+  const isProduction = process.env.NODE_ENV === "production";
+  const monthlyPlanId = isProduction
+    ? "d46c71a7-61dc-4dc8-b53d-9a73d0204c28"
+    : "a02153cd-c49d-49ae-8be6-464296a39a23";
+  const yearlyPlanId = isProduction
+    ? "6fb24b68-09e0-42c4-b090-f0e03cb7de56"
+    : "f3073c34-8b4d-40b7-8123-2f8cbacbc609";
+
+  return (
+    <div className="space-y-6">
+      <DialogHeader>
+        <DialogTitle>Upgrade to Pro</DialogTitle>
+        <DialogDescription>
+          Unlock all features and remove limits.
+        </DialogDescription>
+      </DialogHeader>
+
+      <PlanOption
+        planId={monthlyPlanId}
+        title="Monthly"
+        priceAmount={1900}
+        intervalLabel="Per Month"
+        buttonVariant="outline"
+        isLoading={loadingPlanId === monthlyPlanId}
+        onCheckout={onCheckout}
+      />
+      <PlanOption
+        planId={yearlyPlanId}
+        title="Yearly"
+        priceAmount={9900}
+        intervalLabel="Per Year"
+        badge="Best Value • 20% off"
+        buttonVariant="destructive"
+        isLoading={loadingPlanId === yearlyPlanId}
+        onCheckout={onCheckout}
+      />
+
+      <div className="space-y-3 text-left rounded-md bg-muted/30 p-4">
+        <p className="font-medium text-sm text-muted-foreground">
+          Pro Features included:
+        </p>
+        <ul className="grid grid-cols-1 md:grid-cols-2 gap-2">
+          {featureList.map((feature) => (
+            <li key={feature} className="flex items-center gap-2 text-sm">
+              <CheckCircle2 className="size-4 text-primary shrink-0" />
+              <span>{feature}</span>
+            </li>
+          ))}
+        </ul>
+      </div>
+    </div>
+  );
+}
 
 export default function ProfileSettingsPage() {
   // @ts-ignore
@@ -56,6 +234,47 @@ export default function ProfileSettingsPage() {
   const [deleteError, setDeleteError] = useState<string | null>(null);
   const [signOutLoading, setSignOutLoading] = useState(false);
   const [subscriptionOpen, setSubscriptionOpen] = useState(false);
+  const [loadingPlanId, setLoadingPlanId] = useState<string | null>(null);
+  const [checkoutInstance, setCheckoutInstance] = useState<any>(null);
+  // @ts-ignore
+  const createCheckoutLink = useAction(api.billing.createCheckoutLink);
+
+  useEffect(() => {
+    return () => {
+      if (checkoutInstance) {
+        checkoutInstance.close();
+      }
+    };
+  }, [checkoutInstance]);
+
+  const handleCheckout = async (planId: string) => {
+    setLoadingPlanId(planId);
+    try {
+      const checkoutUrl = await createCheckoutLink({ productId: planId });
+      const checkout = await PolarEmbedCheckout.create(checkoutUrl, "light");
+
+      setCheckoutInstance(checkout);
+      setSubscriptionOpen(false);
+
+      checkout.addEventListener("success", (event: any) => {
+        if (!event.detail.redirect) {
+          toast.success(
+            "Welcome to Pro! Your subscription has been activated."
+          );
+        }
+      });
+
+      checkout.addEventListener("close", (event: any) => {
+        setCheckoutInstance(null);
+      });
+    } catch (error) {
+      console.error("Failed to open checkout", error);
+      toast.error("Failed to start checkout. Please try again.");
+    } finally {
+      setLoadingPlanId(null);
+    }
+  };
+
   const { theme, setTheme } = useTheme();
   const router = useRouter();
 
@@ -437,13 +656,12 @@ export default function ProfileSettingsPage() {
           {useQuery(api.billing.userHasPremium, {}) ? (
             <>
               <Badge>Pro</Badge>
-              <Button
-                size="sm"
-                variant="link"
-                onClick={() => setSubscriptionOpen(!subscriptionOpen)}
+              <CustomerPortalButton
+                className={cn(buttonVariants({ variant: "link", size: "sm" }))}
               >
                 Manage
-              </Button>
+                <ExternalLink />
+              </CustomerPortalButton>
             </>
           ) : (
             <>
@@ -460,7 +678,14 @@ export default function ProfileSettingsPage() {
         </div>
       </div>
 
-      <SubscriptionSection open={subscriptionOpen} />
+      <Dialog open={subscriptionOpen} onOpenChange={setSubscriptionOpen}>
+        <DialogContent className="max-w-3xl">
+          <SubscriptionSection
+            onCheckout={handleCheckout}
+            loadingPlanId={loadingPlanId}
+          />
+        </DialogContent>
+      </Dialog>
 
       <Dialog open={resetOpen} onOpenChange={setResetOpen}>
         <DialogContent>
