+++ apps/web/app/global-error.tsx
@@ -13,7 +13,7 @@ export default function GlobalError({
   return (
     // global-error must include html and body tags
     <html>
-      <body>
+      <body className="flex justify-center items-center flex-col min-h-screen gap-4">
         <h2>Something went wrong!</h2>
         <button onClick={() => reset()} className={cn(buttonVariants())}>
           Try again
